+++ processes/hathiTrust.py
@@ -13,8 +13,8 @@ from model import Record
 class HathiTrustProcess(CoreProcess):
     HATHI_RIGHTS_SKIPS = ['ic', 'icus', 'ic-world', 'und']
 
-    def __init__(self, process, customFile, ingestPeriod):
-        super(HathiTrustProcess, self).__init__(process, customFile, ingestPeriod)
+    def __init__(self, *args):
+        super(HathiTrustProcess, self).__init__(*args[:3])
         self.generateEngine()
         self.createSession()