+++ processes/sfrCluster.py
@@ -8,8 +8,8 @@ from model import Record
 
 
 class ClusterProcess(CoreProcess):
-    def __init__(self, process, customFile, ingestPeriod):
-        super(ClusterProcess, self).__init__(process, customFile, ingestPeriod)
+    def __init__(self, *args):
+        super(ClusterProcess, self).__init__(*args[:3])
 
         # PostgreSQL Connection
         self.generateEngine()